Common: Move SocketAddress function into the class
Moved socketaddress_from_string into SocketAddress.from_string
