--- monkey/common/types/__init__.py
@@ -1,1 +1,0 @@
-from .types import AgentID, HardwareID, MachineID, SocketAddress
