--- monkey/common/types/utils/__init__.py
@@ -1,1 +1,0 @@
-from .socket_address import socketaddress_from_string
