+++ monkey/tests/unit_tests/common/test_socket_address.py
@@ -1,7 +1,6 @@
 import pytest
 
 from common.types import SocketAddress
-from common.types.utils import socketaddress_from_string
 
 GOOD_IP = "192.168.1.1"
 BAD_IP = "192.168.1.999"
@@ -9,10 +8,10 @@ GOOD_PORT = 1234
 BAD_PORT = 99999
 
 
-def test_socketaddress_from_string():
+def test_socket_address__from_string():
     expected = SocketAddress(ip=GOOD_IP, port=GOOD_PORT)
 
-    address = socketaddress_from_string(f"{GOOD_IP}:{GOOD_PORT}")
+    address = SocketAddress.from_string(f"{GOOD_IP}:{GOOD_PORT}")
 
     assert address == expected
 
@@ -30,6 +29,6 @@ def test_socketaddress_from_string():
         f"{GOOD_IP}:{BAD_PORT}",
     ],
 )
-def test_socketaddress_from_string__raises(bad_address: str):
+def test_socket_address__from_string_raises(bad_address: str):
     with pytest.raises(ValueError):
-        socketaddress_from_string(bad_address)
+        SocketAddress.from_string(bad_address)
